add distributed oot registration test
